% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lengths.R
\name{add_lengths}
\alias{add_lengths}
\title{add_lengths}
\usage{
add_lengths(data.df, lengths.df)
}
\arguments{
\item{data.df}{a dataframe with gene_id (must match gene_id in gene coordinates) and either expression data or gene sets}

\item{lengths.df}{a dataframe with gene_id and gene lengths, either the package provided H. sapiens GRCh38.p14 Ensembl Release 110 or user provided with input_lengths()}
}
\value{
returns a dataframe with gene lengths appended
}
\description{
add_lengths
}
\examples{
data(example_expression_data)
data(example_lengths)
data.df <- add_lengths(data.df = example_expression_data,
                       lengths.df = example_lengths)
}
