% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_bar_plot.R
\name{binomial_bar_plot}
\alias{binomial_bar_plot}
\title{binomial_bar_plot()}
\usage{
binomial_bar_plot(binomial.df, by_sample = FALSE, categorical = FALSE)
}
\arguments{
\item{binomial.df}{a dataframe output from binomial_test()}

\item{by_sample}{should data be grouped by sample?}

\item{categorical}{is group categorical?}
}
\value{
returns a gene size enrichment bar plot
}
\description{
binomial_bar_plot()
}
\examples{
# quantitative feature
data(example_quantitative_feature_binomial_test_output)
binomial_bar_plot(
  binomial.df = example_quantitative_feature_binomial_test_output)

# quantitative feature by sample
data(example_quantitative_feature_by_sample_binomial_test_output)
binomial_bar_plot(
  binomial.df = example_quantitative_feature_by_sample_binomial_test_output,
  by_sample = TRUE)

# categorical feature
data(example_categorical_binomial_test_output)
binomial_bar_plot(
  binomial.df = example_categorical_binomial_test_output, categorical = TRUE)
}
