% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_tile_plot.R
\name{binomial_tile_plot}
\alias{binomial_tile_plot}
\title{binomial_tile_plot()}
\usage{
binomial_tile_plot(binomial.df, by_sample = FALSE, categorical = FALSE)
}
\arguments{
\item{binomial.df}{a dataframe output from binomial_test()}

\item{by_sample}{should data be grouped by sample?}

\item{categorical}{is group categorical?}
}
\value{
returns a gene size enrichment tile plot
}
\description{
This is a function named 'binomial_tile_plot'
which generates a binomial gene size enrichment/depletion tile plot
}
\examples{
data(example_quantitative_feature_binomial_test_output)
binomial_tile_plot(
  binomial.df = example_quantitative_feature_binomial_test_output)

data(example_quantitative_feature_by_sample_binomial_test_output)
binomial_tile_plot(
  binomial.df = example_quantitative_feature_by_sample_binomial_test_output,
  by_sample = TRUE)

data(example_categorical_binomial_test_output)
binomial_tile_plot(
  binomial.df = example_categorical_binomial_test_output,
  categorical = TRUE)
}
