% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_categorical_binomial_test_input.R
\docType{data}
\name{example_categorical_binomial_test_input}
\alias{example_categorical_binomial_test_input}
\title{Example categorical binomial test input}
\format{
A data frame with 10 rows and 7 variables:
\describe{
\item{gene_id}{character, ensembl gene id}
\item{group}{character, categorical variable}
\item{length}{numeric, gene length}
...
}
}
\usage{
example_categorical_binomial_test_input
}
\description{
An example dataframe of ensembl gene_id, categorical variable, and gene lengths
}
\keyword{datasets}
