% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_expression_data.R
\docType{data}
\name{example_expression_data}
\alias{example_expression_data}
\title{Example expression data}
\format{
A data frame with 10 rows and 2 variables:
\describe{
\item{gene_id}{character, ensembl gene id}
\item{log2FC}{numeric, expression}
...
}
}
\usage{
example_expression_data
}
\description{
An example dataframe of ensembl gene_id and expression values (here log2FC, but could be any numerical value)
}
\keyword{datasets}
