% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/example_quantitative_feature_binomial_test_input.R
\docType{data}
\name{example_quantitative_feature_binomial_test_input}
\alias{example_quantitative_feature_binomial_test_input}
\title{Example quantitative feature binomial test input}
\format{
A data frame with 10 rows and 3 variables:
\describe{
\item{gene_id}{character, ensembl gene id}
\item{log2FC}{numeric, expression}
\item{length}{numeric, gene length}
...
}
}
\usage{
example_quantitative_feature_binomial_test_input
}
\description{
An example dataframe of ensembl gene_id, expression values (here log2FC, but could be any numerical value), and gene lengths
}
\keyword{datasets}
