% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/example_quantitative_feature_binomial_test_output.R
\docType{data}
\name{example_quantitative_feature_binomial_test_output}
\alias{example_quantitative_feature_binomial_test_output}
\title{Example quantitative feature binomial test output}
\format{
A data frame with 20 rows and 7 variables:
\describe{
\item{length_range}{ord, gene length range}
\item{feature_range}{ord, quantitative variable range}
\item{n}{numeric, number of genes}
\item{total_trials}{numeric, number of genes in the group and length_range}
\item{p_val}{numeric, two-sided binomial test p-value}
\item{p_adj}{numeric, BH adjusted p-value}
\item{sign}{character, gene-size enrichement or depletion}
...
}
}
\usage{
example_quantitative_feature_binomial_test_output
}
\description{
An example dataframe of length_range, feature_range, n, total_trials, p_val, p_adj, and sign
}
\keyword{datasets}
