% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/example_quantitative_feature_by_sample_binomial_test_input.R
\docType{data}
\name{example_quantitative_feature_by_sample_binomial_test_input}
\alias{example_quantitative_feature_by_sample_binomial_test_input}
\title{Example quantitative feature by sample binomial test input}
\format{
A data frame with 3742 rows and 6 variables:
\describe{
\item{sample1}{numeric, gene expression}
\item{sample2}{numeric, gene expression}
\item{sample3}{numeric, gene expression}
\item{sample4}{numeric, gene expression}
\item{gene_id}{character, ensembl gene id}
\item{length}{numeric, gene length}
...
}
}
\usage{
example_quantitative_feature_by_sample_binomial_test_input
}
\description{
An example dataframe of expression values (here TPM, but could be any quantitative variable), ensembl gene_id, and gene lengths
}
\keyword{datasets}
