% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_lengths.R
\name{gene_lengths}
\alias{gene_lengths}
\title{gene_lengths: a function that allows users to input their own ensembl gene
list to calculate gene lengths}
\usage{
gene_lengths(filepath)
}
\arguments{
\item{filepath}{a filepath to a tab-delimited file with chromosome,
start position, end positions, and ensembl gene id (in that order)
for each gene}
}
\value{
returns dataframe with gene lengths for user provided genes
}
\description{
gene_lengths: a function that allows users to input their own ensembl gene
list to calculate gene lengths
}
\examples{
lengths.df <- gene_lengths(
  file = system.file(
    "extdata",
    "example_gene_coordinates.tsv",
    package = "sizeR"))
}
