% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesizeR_binomial_test.R
\name{genesizeR_binomial_test}
\alias{genesizeR_binomial_test}
\title{genesizeR_binomial_test}
\usage{
genesizeR_binomial_test(
  data.df,
  by_sample = FALSE,
  categorical = FALSE,
  feature_name = NULL,
  alpha = 0.005
)
}
\arguments{
\item{data.df}{a dataframe containing gene_id, length, and either gene set or expression data}

\item{by_sample}{should data be grouped by sample?}

\item{categorical}{is group categorical?}

\item{feature_name}{the exact feature name in the data (e.g. log2FC)}

\item{alpha}{the p-value cutoff (default p < 0.005)}
}
\value{
returns a dataframe with estimates for gene size enrichment by binomial test
}
\description{
genesizeR_binomial_test
}
\examples{
# quantitative feature
data(example_quantitative_feature_binomial_test_input)
data.df <- genesizeR_binomial_test(
  data.df = example_quantitative_feature_binomial_test_input,
  feature_name = "log2FC")

# quantitative feature by sample
data(example_quantitative_feature_by_sample_binomial_test_input)
data.df <- genesizeR_binomial_test(
  data.df = example_quantitative_feature_by_sample_binomial_test_input,
  by_sample = TRUE)

# categorical feature
data(example_categorical_binomial_test_input)
data.df <- genesizeR_binomial_test(
  data.df = example_categorical_binomial_test_input, categorical = TRUE)
}
