% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesizeR_gene_lengths.R
\name{genesizeR_gene_lengths}
\alias{genesizeR_gene_lengths}
\title{genesizeR_gene_lengths: a function that allows users to input their own
ensembl gene list to calculate gene lengths}
\usage{
genesizeR_gene_lengths(filepath, gene_id, delim)
}
\arguments{
\item{filepath}{a filepath to a tab-delimited file with chromosome,
start position, end positions, and ensembl gene id (in that order)}

\item{gene_id}{name of gene_id column (e.g. \verb{Gene stable ID}).
This must match the gene_id name in data.}

\item{delim}{delimiter in file}
}
\value{
returns dataframe with gene lengths for user provided genes
}
\description{
genesizeR_gene_lengths: a function that allows users to input their own
ensembl gene list to calculate gene lengths
}
\examples{
lengths.df <- genesizeR_gene_lengths(
  file = system.file(
    "extdata",
    "example_gene_coordinates.tsv",
    package = "genesizeR"),
  gene_id = "Gene stable ID",
  delim = "\t")
}
