% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesizeR_plot.R
\name{genesizeR_plot}
\alias{genesizeR_plot}
\title{genesizeR_plot()}
\usage{
genesizeR_plot(
  data.df,
  by_sample = FALSE,
  categorical = FALSE,
  quantiles = TRUE,
  type
)
}
\arguments{
\item{data.df}{a dataframe output from binomial_test()}

\item{by_sample}{should data be grouped by sample?}

\item{categorical}{is group categorical?}

\item{quantiles}{should x and y values be quantiles?}

\item{type}{geom type (i.e. bar, tile, line)}
}
\value{
returns a genesizeR plot
}
\description{
genesizeR_plot()
}
\examples{

###################
# genesizeR bar plots #
###################

# quantitative feature
data(example_quantitative_feature_binomial_test_output)
genesizeR_plot(
  data.df = example_quantitative_feature_binomial_test_output,
  quantiles = FALSE,
  type = "bar")

# quantitative feature by sample
data(example_quantitative_feature_by_sample_binomial_test_output)
genesizeR_plot(
  data.df = example_quantitative_feature_by_sample_binomial_test_output,
  by_sample = TRUE,
  quantiles = FALSE,
  type = "bar")

# categorical feature
data(example_categorical_binomial_test_output)
genesizeR_plot(
  data.df = example_categorical_binomial_test_output, categorical = TRUE,
  quantiles = FALSE,
  type = "bar")

####################
# genesizeR tile plots #
####################

data(example_quantitative_feature_binomial_test_output)
genesizeR_plot(
  data.df = example_quantitative_feature_binomial_test_output,
  quantiles = FALSE,
  type = "tile")

data(example_quantitative_feature_by_sample_binomial_test_output)
genesizeR_plot(
  data.df = example_quantitative_feature_by_sample_binomial_test_output,
  by_sample = TRUE,
  quantiles = FALSE,
  type = "tile")

data(example_categorical_binomial_test_output)
genesizeR_plot(
  data.df = example_categorical_binomial_test_output,
  categorical = TRUE,
  quantiles = FALSE,
  type = "tile")

###################
# genesizeR line plot #
###################

data(example_quantitative_feature_by_sample_binomial_test_input)
genesizeR_plot(
  data.df = example_quantitative_feature_by_sample_binomial_test_input,
  quantiles = FALSE,
  type = "line")
}
