% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeR_input.R
\name{sizeR_input}
\alias{sizeR_input}
\title{sizeR_input: a function to input gene expression or gene set data}
\usage{
sizeR_input(filepath, gene_id, delim)
}
\arguments{
\item{filepath}{a filepath to a tab-delimited file with gene_id and
expression data or gene set data}

\item{gene_id}{name of gene_id column (e.g. \verb{Gene stable ID}).
This must match the gene_id name in data.}

\item{delim}{delimiter in file (e.g. "\t", or ",")}
}
\value{
returns dataframe with gene_id and expression or gene set data
}
\description{
sizeR_input: a function to input gene expression or gene set data
}
\examples{
data.df <- sizeR_input(
  file = system.file(
    "extdata",
    "example_gene_coordinates.tsv",
    package = "sizeR"),
  gene_id = "Gene stable ID",
  delim = "\t")
}
