% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeR_line_plot.R
\name{sizeR_line_plot}
\alias{sizeR_line_plot}
\title{sizeR_line_plot()}
\usage{
sizeR_line_plot(data.df)
}
\arguments{
\item{data.df}{a dataframe containing gene length and expression values}
}
\value{
returns a gene size versus expression line plot
}
\description{
This is a function named 'sizeR_line_plot'
which generates a gene size versus expression line plot
}
\examples{
data(example_quantitative_feature_by_sample_binomial_test_input)
sizeR_line_plot(
  data.df = example_quantitative_feature_by_sample_binomial_test_input)
}
