% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeR_plot.R
\name{sizeR_plot}
\alias{sizeR_plot}
\title{sizeR_plot()}
\usage{
sizeR_plot(
  data.df,
  by_sample = FALSE,
  categorical = FALSE,
  quantiles = TRUE,
  type
)
}
\arguments{
\item{data.df}{a dataframe output from binomial_test()}

\item{by_sample}{should data be grouped by sample?}

\item{categorical}{is group categorical?}

\item{quantiles}{should x and y values be quantiles?}

\item{type}{geom type (i.e. bar, tile, line)}
}
\value{
returns a sizeR plot
}
\description{
sizeR_plot()
}
\examples{

###################
# sizeR bar plots #
###################

# quantitative feature
data(example_quantitative_feature_binomial_test_output)
sizeR_plot(
  data.df = example_quantitative_feature_binomial_test_output,
  quantiles = FALSE,
  type = "bar")

# quantitative feature by sample
data(example_quantitative_feature_by_sample_binomial_test_output)
sizeR_plot(
  data.df = example_quantitative_feature_by_sample_binomial_test_output,
  by_sample = TRUE,
  quantiles = FALSE,
  type = "bar")

# categorical feature
data(example_categorical_binomial_test_output)
sizeR_plot(
  data.df = example_categorical_binomial_test_output, categorical = TRUE,
  quantiles = FALSE,
  type = "bar")

####################
# sizeR tile plots #
####################

data(example_quantitative_feature_binomial_test_output)
sizeR_plot(
  data.df = example_quantitative_feature_binomial_test_output,
  quantiles = FALSE,
  type = "tile")

data(example_quantitative_feature_by_sample_binomial_test_output)
sizeR_plot(
  data.df = example_quantitative_feature_by_sample_binomial_test_output,
  by_sample = TRUE,
  quantiles = FALSE,
  type = "tile")

data(example_categorical_binomial_test_output)
sizeR_plot(
  data.df = example_categorical_binomial_test_output,
  categorical = TRUE,
  quantiles = FALSE,
  type = "tile")

###################
# sizeR line plot #
###################

data(example_quantitative_feature_by_sample_binomial_test_input)
sizeR_plot(
  data.df = example_quantitative_feature_by_sample_binomial_test_input,
  quantiles = FALSE,
  type = "line") +
  scale_color_manual(values = c("condition1.r1" = "firebrick", "condition1.r2" = "firebrick", "condition2.r1" = "black", "condition2.r2" = "black"))
}
