% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_lengths.R
\docType{data}
\name{example_lengths}
\alias{example_lengths}
\title{Example lengths}
\format{
A data frame with 10 rows and 5 variables:
\describe{
\item{Chromosome/scaffold name}{character, chromosome}
\item{Gene start (bp)}{numeric, start position}
\item{Gene end (bp)}{numeric, end position}
\item{Gene stable ID}{character, ensembl gene id}
\item{length}{numeric, gene length}
...
}
}
\usage{
example_lengths
}
\description{
An example dataframe of gene coordinates and gene lengths
}
\keyword{datasets}
